requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.0
python-decouple==3.3
django-rest-swagger==2.2.0